% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kandy-package.R
\docType{package}
\name{kandy-package}
\alias{kandy}
\alias{kandy-package}
\title{kandy: Multivariate-Normal probability density function}
\description{
A collection of function for computing the density for a multivariate Gaussian distibution. Several implementations are provided to examplify the different techniques for optimizing and accelerating R codes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/blondykayembe/kandy}
  \item \url{https://blondykayembe.github.io/kandy/}
  \item Report bugs at \url{https://github.com/blondykayembe/kandy/issues}
}

}
\author{
\strong{Maintainer}: Blondy Kayembe Mulumba \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
