% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kandy-package.R
\docType{package}
\name{kandy-package}
\alias{kandy}
\alias{kandy-package}
\title{kandy: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Blondy KM \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
