% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kandy-package.R
\docType{package}
\name{kandy-package}
\alias{kandy}
\alias{kandy-package}
\title{kandy: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/blondykayembe/kandy}
  \item \url{https://blondykayembe.github.io/kandy/}
  \item Report bugs at \url{https://github.com/blondykayembe/kandy/issues !!!OVERTHROW !!!}
}

}
\author{
\strong{Maintainer}: Blondy KM \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
