% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mnvpdfsmart.R,
%   R/mvnpdfRox.R, R/mvnpdf_invC.R, R/mvnpdfoptim.R, R/mvnpdfsmart_par.R,
%   R/mvnpdfsmart_sleepy.R, R/mvnpdfsmart_sleepy_par.R
\name{invC}
\alias{invC}
\alias{mvnpdfoptimC}
\alias{mvnpdfsmartC}
\alias{timesTwo}
\alias{mvnpdfsmart}
\alias{mvnpdf}
\alias{mvnpdf_invC}
\alias{mvnpdfoptim}
\alias{mvnpdfsmart_par}
\alias{mvnpdfsmart_sleepy}
\alias{mvnpdfsmart_sleepy_par}
\title{Based on the implementation from Nino Hardt and Dicko Ahmadou
http://gallery.rcpp.org/articles/dmvnorm_arma/
(accessed in August 2014)}
\usage{
invC(M)

mvnpdfoptimC(x, mean, varcovM, Log = TRUE)

mvnpdfsmartC(x, mean, varcovM, Log = TRUE)

timesTwo(x)

mvnpdfsmart(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdf(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdf_invC(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfoptim(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart_par(x, mean = rep(0, nrow(x)), varcovM = diag(nrow(x)), Log = TRUE)

mvnpdfsmart_sleepy(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE
)

mvnpdfsmart_sleepy_par(
  x,
  mean = rep(0, nrow(x)),
  varcovM = diag(nrow(x)),
  Log = TRUE
)
}
\arguments{
\item{M}{matrix}

\item{x}{x is a matrix containing the observations in columns}

\item{mean}{mean vector of the multivariate gaussian distribution}

\item{varcovM}{variance covariance matrix}

\item{Log}{logical parameter (default = True)}
}
\value{
a list of values
}
\description{
description
}
\details{
details
}
\examples{
X = matrix(c(-0.5, 1.5, 0, 1, -1, 1), nrow = 2)
mvnpdf(X)

}
